pyproject.toml
[project]
name = "comfyui-dream-video-batches"
description = "Provide utilities for batch based video generation workflows (s.a. AnimateDiff and Stable Video Diffusion)."
version = "1.1.0"
license = { file = "LICENSE" }
dependencies = ["imageio", "pilgram", "scipy", "numpy<2.0,>=1.18", "torchvision", "evalidate"]

[project.urls]
Repository = "https://github.com/alt-key-project/comfyui-dream-video-batches"
#  Used by Comfy Registry https://comfyregistry.org

[tool.comfy]
PublisherId = "altkeyproject"
DisplayName = "comfyui-dream-video-batches"
Icon = ""
